td-benchmark: Refactor make the library modular.
1. Make library structure more clearer
   Split lib.rs. Make benchmark function to a separate file.
2. Allow heap, stack profiling functions to be used without `benchmark`
   related dependencies, `td-layout`, `x86`, `scroll`.
